--- components/PFP/index.tsx
@@ -1,3 +1,0 @@
-import PFP from "./PFP"
-
-export default PFP
